Moved Path into Types library
