+++ sql/cell_value.cc
@@ -20,8 +20,8 @@
 #include "mldb/base/parse_context.h"
 #include "mldb/compiler/compiler.h"
 #include "interval.h"
-#include "path.h"
 #include <math.h>
+#include "mldb/types/path.h"
 #include "mldb/ext/s2/s2.h"
 #include "mldb/utils/possibly_dynamic_buffer.h"
 #include "mldb/base/less.h"
